+++ components/selfupdater/api/src/main/kotlin/com/flipperdevices/selfupdater/api/SelfUpdate.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.selfupdater.api
+
+data class SelfUpdate(
+    val version: String,
+    val downloadUrl: String,
+    val name: String,
+)