+++ components/selfupdater/api/src/main/kotlin/com/flipperdevices/selfupdater/api/SelfUpdateParserApi.kt
@@ -1,0 +1,6 @@
+package com.flipperdevices.selfupdater.api
+
+interface SelfUpdateParserApi {
+    fun getName(): String
+    suspend fun getLastUpdate(): SelfUpdate?
+}